refactor(swap): extract default-destination seeding into useDefaultSwapDestination
